Add Service Worker
Added a service worker to the edge worker. Instead of letting the browser handle the service worker, we're now using the edge worker to handle the service worker. The worker will  intercept the redirect response and fetch the content itself. It will update the URl seperately using histpory api. Essentially, the service worker is now a proxy for the edge worker. content will be proxied and avoid aredirect.